slave/requirements.txt
requests
socketIO-client-2
